+++ migrations/2_registry.js
@@ -1,0 +1,5 @@
+const Registry = artifacts.require('Registry');
+
+module.exports = function(deployer) {
+  deployer.deploy(Registry);
+};