+++ migrations/4_haaveprotocol.js
@@ -1,0 +1,20 @@
+const Registry = artifacts.require('Registry');
+const Handler = artifacts.require('HAaveProtocol');
+const utils = web3.utils;
+const AAVE_LENDING_POOL_CORE = '0x398ec7346dcd622edc5ae82352f02be94c62d119';
+
+module.exports = function(deployer) {
+  deployer
+    .deploy(Handler)
+    .then(function() {
+      return Registry.deployed();
+    })
+    .then(function(instance) {
+      registry = instance;
+      registry.register(AAVE_LENDING_POOL_CORE, Handler.address);
+      return registry.register(
+        Handler.address,
+        utils.asciiToHex('Aave Protocol')
+      );
+    });
+};